feat: change detailed page to modal, add params hook, pagination only after receiving products
