+++ src/components/cardList/CardList.module.scss
@@ -9,7 +9,6 @@
   align-content: center;
   gap: 10px;
   padding: 50px 6px 100px 6px;
-  background-color: $grey;
   border-radius: 6px;
   min-height: 70vh;
 }
