+++ src/components/detailedCard/DetailedCard.module.scss
@@ -9,9 +9,12 @@
   border: 1px solid $black;
   border-radius: 6px;
   padding: 5px;
+  margin-right: 5px;
   box-shadow: $black-shadow;
   text-align: center;
-  width: 300px;
+  background-color: $white;
+  position: relative;
+  height: fit-content;
 }
 
 .closeButton {
