+++ src/components/detailedCard/DetailedCard.tsx
@@ -1,23 +1,16 @@
-import { useSearchParams } from 'react-router-dom';
-
-import { CardProps } from 'src/components/card/Card';
+import { Product } from 'src/interfaces/product';
 
 import styles from 'src/components/detailedCard/DetailedCard.module.scss';
 
+interface DetailedCardProps {
+  product: Product;
+  closeDetailedPage: (e: React.MouseEvent) => void;
+}
+
 export default function DetailedCard({
-  product: { description, name, image_url, first_brewed },
-}: CardProps): JSX.Element {
-  const [, setSearchParams] = useSearchParams();
-
-  const closeDetailedPage = (e: React.MouseEvent) => {
-    if (e.target === e.currentTarget) {
-      setSearchParams((searchParams) => {
-        searchParams.delete('productId');
-        return searchParams;
-      });
-    }
-  };
-
+  product: { description, tagline, name, image_url, first_brewed },
+  closeDetailedPage,
+}: DetailedCardProps): JSX.Element {
   return (
     <div className={styles.detailedCard}>
       <button className={styles.closeButton} onClick={closeDetailedPage}>
@@ -25,6 +18,7 @@ export default function DetailedCard({
       </button>
       <h3 className={styles.cardHeading}>{name}</h3>
       <img src={image_url} alt={`${name} image`} className={styles.cardImage} />
+      <p className={styles.cardText}>{tagline}</p>
       <p className={styles.cardText}>{description}</p>
       <p className={styles.cardText}>{first_brewed}</p>
     </div>
