+++ src/pages/DetailedPage/DetailedPage.module.scss
@@ -1,12 +1,19 @@
+@import 'src/styles/variables.scss';
+
 .container {
-  padding-top: 50px;
-  display: flex;
+  display: grid;
   justify-content: center;
-  width: 320px;
+  width: 100%;
+  height: 100%;
+  left: 0;
+  top: 0;
+  position: fixed;
+  grid-template-columns: 2fr 1fr;
   margin: 0 auto;
-  div {
-    position: fixed;
-  }
+  z-index: 2;
+  background-color: $transparent-black;
+  align-items: center;
+  overflow: hidden;
 }
 
 @import 'src/styles/media.scss';
